Extract keyring mocks
