+++ packages/ui/test/pages/Profile.test.tsx
@@ -6,10 +6,10 @@ import { TypeRegistry } from '@polkadot/types'
 import metaStatic from '@polkadot/metadata/static'
 import { Metadata } from '@polkadot/metadata'
 import keyring from '@polkadot/ui-keyring'
-import { KeyringJson, KeyringStore } from '@polkadot/ui-keyring/types'
 import { Profile } from '../../src/pages/Profile/Profile'
 import { SubstrateContext } from '../../src/providers/substrate/context'
 import { State, types } from '../../src/providers/substrate/reducer'
+import { aliceSigner, MemoryStore } from '../mocks/keyring'
 
 export function createApiWithAugmentations(): ApiPromise {
   const registry = new TypeRegistry()
@@ -22,45 +22,6 @@ export function createApiWithAugmentations(): ApiPromise {
   api.injectMetadata(metadata, true)
 
   return api
-}
-
-type AccountsMap = Record<string, KeyringJson>
-
-export class MemoryStore implements KeyringStore {
-  private accounts: AccountsMap = {}
-
-  all(cb: (key: string, value: KeyringJson) => void): void {
-    Object.keys(this.accounts).forEach((accountsKey) => cb(accountsKey, this.accounts[accountsKey]))
-  }
-
-  get(key: string, cb: (value: KeyringJson) => void): void {
-    cb(this.accounts[key])
-  }
-
-  remove(key: string, cb: (() => void) | undefined): void {
-    delete this.accounts[key]
-
-    if (cb) {
-      cb()
-    }
-  }
-
-  set(key: string, value: KeyringJson, cb: (() => void) | undefined): void {
-    this.accounts[key] = value
-
-    if (cb) {
-      cb()
-    }
-  }
-}
-
-import { Keyring } from '@polkadot/keyring'
-import { KeyringPair } from '@polkadot/keyring/types'
-
-export function aliceSigner(): KeyringPair {
-  const keyring = new Keyring({ type: 'sr25519' })
-
-  return keyring.addFromUri('//Alice')
 }
 
 describe('UI: Profile', () => {
